Delete ride from user account functionality added. Job to delete users from queue added. update notification API to show specific waiting time

+++ Virtual_Queue/src/main/java/com/virtual/queue/service/NotificationService.java
@@ -5,8 +5,8 @@ import java.util.List;
 import com.virtual.queue.beans.UserQueueInfo;
 
 public interface NotificationService {
-	public List<UserQueueInfo> pullNotInfo(Integer rideId);
+	public List<UserQueueInfo> pullNotInfo(long rideId);
 	public List<UserQueueInfo> pullAllNotInfo();
-	public void notifyUser(Integer rideId) throws Exception;
+	public void notifyUser(long rideId) throws Exception;
 	public void notifyAllUsers() throws Exception;
 }
